Update Invoice page UI and fix css of invoice form
