+++ src/App.css
@@ -105,6 +105,7 @@ justify-content: center;
 .Component-kanban-card-row{
   margin-top: 32px;
   margin-left: 10px;
+  margin-right: 10px;
 }
 .Component-upload-progress{
   margin-top: 5px;
