+++ src/pages/Dashboard/index.tsx
@@ -1,14 +1,10 @@
 import {UserContext} from "../../stores/user";
-import { Header } from "../../components/Header";
 import Box from "@mui/material/Box";
 import Grid from '@mui/material/Grid';
 import CreateButton from "../../components/CreateButton";
-import { useNavigate } from "react-router-dom";
 import { BankForm } from "../../components/BankForm";
 import {useContext, useEffect, useState} from "react";
 import {getAllReimbursement, getMyReimbursement, ReimbursementRead} from "../../database/reimbursement";
-import {createUser} from "../../auth/types";
-import {logoutSession} from "../../auth/session";
 import { KanbanBoard } from "../../components/KanbanBoard";
 import ReimbursementCard from "../../components/Card/ReimbursementCard";
 
@@ -16,20 +12,6 @@ import ReimbursementCard from "../../components/Card/ReimbursementCard";
 
 export default function Dashboard() {
   const { user, setUser: setUserStore } = useContext(UserContext);
-  const navigate = useNavigate();
-
-  const handleLogout = async () => {
-    console.log("logout");
-
-    setUserStore(createUser(null));
-    await logoutSession();
-    navigate("/login");
-  };
-
-  const handleAnalytics = () => {
-    navigate("/analytics");
-  }
-
   const [reimbursement, setReimbursement] = useState<ReimbursementRead[]>([]);
 
   useEffect(() => {
@@ -49,19 +31,13 @@ export default function Dashboard() {
           <Grid container spacing={2}>
             <Grid item xs={12} md={3}>
               <CreateButton
-                link="#/invoice"
-                title="Create invoice"
-              ></CreateButton>
-            </Grid>
-            <Grid item xs={12} md={3}>
-              <CreateButton
                 link="#/reimbursement"
                 title="Create reimbursement"
               ></CreateButton>
             </Grid>
           </Grid>
+          <KanbanBoard user={user} statusContainers={reimbursement}  Card={ReimbursementCard}/>
         </Box>
-        <KanbanBoard user={user} statusContainers={reimbursement}  Card={ReimbursementCard}/>
         <BankForm />
       </div>
     </>
