+++ src/routes/RouterProvider.tsx
@@ -10,7 +10,7 @@ import InvoiceGeneration from "../pages/InvoiceGeneration";
 import ReimbursementGeneration from "../pages/ReimbursementGeneration";
 import {Analytics} from "../pages/Analytics";
 import {Header} from "../components/Header";
-import {Invoices} from "../pages/Invoices";
+import Invoices from "../pages/Invoices";
 
 function RouterProvider() {
   return <>
